Move the schema writer into the `graph` package
